Updated db for pull
- Server requries db.js
- db.js connects to database and retrieves models
- models.js contains schemas for all db data
- renamed app-db.js to db.js
- Updated readme with usage instructions for Windows
- Made whitespace and tabs for DB so it looks proper
- Updated server to have models from the database
- added PasswordHash and twoFactorMethod to User
